Switch to Lavalink-based voice commands as main voice commands.
