+++ src/commands/mod.rs
@@ -2,6 +2,6 @@ pub mod general_group;
 pub mod help;
 pub mod server_mod_group;
 pub mod emoji_group;
-pub mod voice_group;
+pub mod legacy_voice_group;
 pub mod checks;
-pub mod voicelavalink_group;
+pub mod voice_group;
