+++ src/system_fn.rs
@@ -24,7 +24,7 @@ use crate::wrapper::check_msg;
 // This additional information will also only be shown if the LOG level is set to `debug`
 #[instrument]
 pub async fn before(_ctx: &Context, msg: &Message, command_name: &str) -> bool {
-    debug!("Got command '{}' by user '{}'", command_name, msg.author.name);
+    info!("Got command '{}' by user '{}'", command_name, msg.author.name);
     true
 }
